fix(webapp): move API client types to shared
