+++ apps/shared/src/types/index.ts
@@ -79,6 +79,7 @@ export interface ContractValuationPayload {
 }
 export * from "./risk";
 export * from "./pagination";
+export * from "./api";
 export type {
   GameEvent,
   GameEventType,
